test: merge each test directory into one file
Each of the three test directories under test/ now has one .ts file:
test/pcre2-compatibility.test.ts, test/pcre2-extensions.test.ts, and
test/pcre2-js-divergences.test.ts. Every describe and it block moves
across unchanged, grouped by source file under a wrapping describe.
Shared imports are merged into one import block per file.

Update the one console.log path reference in build-corpus.mjs to
match.
